PGFReader
Ignore-this: 6a71a754279cbb1cfd4371488f38a8c1

darcs-hash:20100420220214-b2669-6c3425880a6ae31632325a8f037e58c3f2ac0adc.gz

+++ src/old/java/pgf/reader/FloatLiteral.java
@@ -1,9 +1,9 @@
 package reader;
 
 public class FloatLiteral extends Literal {
-      float value;
+      double value;
        
-	  public FloatLiteral(float _value) {value = _value; }
+	  public FloatLiteral(double _value) {value = _value; }
 
 	  public <R,A> R accept(reader.Literal.Visitor<R,A> v, A arg) { return v.visit(this, arg); }
